Allow passing impersonated device ID to Appservice and Intent
When creating an impersonating client for an appservice, allow passing in a device ID to use. This will make the requests have `org.matrix.msc3202.device_id` in the query string for the client requests.

Appservice `intentsConfig` can be provided with a `botDeviceId`, which will be passed into creating the `Intent` for the bot user only.